making the listing of artifact-stores consistent
- Resolves #8 inconsistent cli params
- artifact-store show is now list
- updated the documentation in commands.md

There was inconsistency in the cli params when listing external
resources and artifact stores. I chose to use list for both of
them because there might be multiple in use and list makes more
sense to me.

There was also a mistake in the docs where it said ls instead of
list, this is fixed as well. Also there was no documentation of
artifact-store command params, which exists now.